+++ modelo/treino_continuo.py
@@ -1,7 +1,3 @@
-import sys
-import os
-sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
-
 import pandas as pd
 import joblib
 from sklearn.ensemble import RandomForestClassifier
