+++ scripts/teste_firebase.py
